% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_roots}
\alias{all_roots}
\title{Finds all roots of a function within an interval}
\usage{
all_roots(
  f,
  interval,
  lower = min(interval),
  upper = max(interval),
  tol = .Machine$double.eps^0.2,
  maxiter = 1000,
  trace = 0,
  n = 1000,
  ...
)
}
\arguments{
\item{f}{The function for which the root is sought..}

\item{interval}{A vector containing the end-points of the interval to be
searched for the root.}

\item{lower}{Lower end point of interval.}

\item{upper}{Upper end point of interval.}

\item{tol}{Desired accuracy of root}

\item{maxiter}{Maximum number of iterations}

\item{trace}{Integer number passed to function uniroot. If positive, tracing information is produced.
Higher values giving more details.}

\item{n}{Number of subintervals in which the root is sought.}

\item{...}{Additional named or unnamed arguments to be passed to f.}
}
\description{
Searches the interval from lower to upper for all roots of a function f with respect
    to its first argument.  Based on rootSolve::uniroot.all.
}
