% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compute_count_function}
\alias{compute_count_function}
\title{Computes count function over time}
\usage{
compute_count_function(events, T_max, N = 5000)
}
\arguments{
\item{events}{List of event times}

\item{T_max}{Maximum time of simulation to evaluate function at}

\item{N}{Number of points to evaluate function at}
}
\value{
Data frame with times and events counts.
}
\description{
Computes count function over time
}
