% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{compute_integral_limits}
\alias{compute_integral_limits}
\title{Computes total integral}
\usage{
compute_integral_limits(
  time,
  T_max,
  T_min,
  mu_fn,
  mu_diff_fn,
  mu_int_fn,
  parameters,
  print_level = 1
)
}
\arguments{
\item{time}{Current time}

\item{T_max}{Maximum time of simulation}

\item{T_min}{Minimum time of simulation}

\item{mu_fn}{Mu function}

\item{mu_diff_fn}{Derivative of mu function}

\item{mu_int_fn}{Integral of mu function}

\item{parameters}{Parameters}

\item{print_level}{Level of logger to print}
}
\value{
Returns integral of function between limits.  Accounts for functions being zero in
some instances.
}
\description{
Computes total integral
}
\examples{
compute_integral_limits(time = 2000, T_max = 2000, T_min = 0,
                        mu_fn = mu_constant, mu_diff_fn = mu_constant,
                        mu_int_fn = mu_int_constant, parameters = list("A" = 5))
compute_integral_limits(time = 2000, T_max = 2000, T_min = 0,
                        mu_fn = mu_sinusoidal, mu_diff_fn = mu_sinusoidal,
                        mu_int_fn = mu_int_sinusoidal,
                        parameters = list("M" = 1, "N" = 1))
}
