% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute value of intensity function}
\usage{
compute_lambda(
  time,
  previous_event_time,
  events,
  kernel,
  parameters,
  mu_fn = mu_none
)
}
\arguments{
\item{time}{Current time.}

\item{previous_event_time}{Time of previous event.}

\item{events}{List of previous events in simulation.}

\item{kernel}{Kernel function}

\item{parameters}{Parameters for the Hawkes Kernel.}

\item{mu_fn}{Exogenous contribution to intensity.}
}
\value{
Message printed to console.
}
\description{
Compute value of intensity function
}
