% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_intensities.R
\name{conditional_intensity}
\alias{conditional_intensity}
\title{Compute conditional intensity of events at time time}
\usage{
conditional_intensity(time, kernel, events, parameters)
}
\arguments{
\item{time}{Current time.}

\item{kernel}{Function describing the kernel.}

\item{events}{Vector of event times.}

\item{parameters}{Parameters of the Hawkes kernel.}
}
\value{
The sum of the kernel evaluated at \code{t} - \code{y}.
}
\description{
Compute conditional intensity of events at time time
}
\examples{
conditional_intensity(time = 1.5, kernel = ray_kernel, events = c(0.5, 1, 1.3),
    parameters = list("alpha" = 1.0, "delta" = 1.0, "delay" = 0))
}
