% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_intensities.R
\name{conditional_intensity_list}
\alias{conditional_intensity_list}
\title{Computes conditional intensity for a list of events and times}
\usage{
conditional_intensity_list(times, kernel, events, parameters)
}
\arguments{
\item{times}{List of current time.}

\item{kernel}{Function describing the kernel.}

\item{events}{Vector of event times.}

\item{parameters}{Parameters of the Hawkes kernel.}
}
\value{
A vector of the sums of the kernel evaluated at \code{t} - \code{y}.
}
\description{
Computes conditional intensity for a list of events and times
}
\examples{
conditional_intensity_list(time = c(1, 1.5), kernel = ray_kernel,
    events = c(0.5, 1, 1.3), parameters = list("alpha" = 1.0, "delta" = 1.0, "delay" = 0))
}
