% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_derivatives.R
\name{exp_derivatives}
\alias{exp_derivatives}
\title{Computes directional derivatives for exponential kernel}
\usage{
exp_derivatives(
  parameters,
  delay = 0,
  events,
  kernel = exp_kernel,
  mu_fn = mu_none,
  mu_diff_fn = mu_diff_none,
  mu_int_fn = mu_int_none,
  print_level = 0
)
}
\arguments{
\item{parameters}{Parameters of the Hawkes kernel.}

\item{delay}{Fixed delay for kernel}

\item{events}{Vector of event times.}

\item{kernel}{Kernel function for Hawkes Process.}

\item{mu_fn}{Function that returns exogenous part of Hawkes Process.}

\item{mu_diff_fn}{Function that returns differential of exogenous part.}

\item{mu_int_fn}{Function that returns integral of exogenous part.}

\item{print_level}{Level at which logger will print}
}
\value{
Returns directional derivatives for exponential kernel.
}
\description{
Computes directional derivatives for exponential kernel
}
\examples{
exp_derivatives(parameters = list("alpha" = 1, "delta" = 1),
    events = c(0, 4, 6, 9), kernel = exp_kernel)
}
