% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{exp_kernel}
\alias{exp_kernel}
\title{Exponential kernel function}
\usage{
exp_kernel(t, parameters)
}
\arguments{
\item{t}{Current time.}

\item{parameters}{Parameters of the Hawkes kernel.}
}
\value{
Value of the exponential kernel at time \code{t}.
}
\description{
One potential kernel function for use in the Hawkes Process model.
}
\examples{
exp_kernel(5, parameters = list("alpha" = 1, "delta" = 1))
exp_kernel(3, parameters = list("alpha" = 1, "delta" = 0.5))
}
