% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{int_exp}
\alias{int_exp}
\title{Integral of exponential kernel function between T_max and 0}
\usage{
int_exp(T_max, events, parameters)
}
\arguments{
\item{T_max}{Maximum time in data.}

\item{events}{Vector of event times.}

\item{parameters}{Parameters of the Hawkes kernel.}
}
\value{
Value of the integral of the exponential kernel at time \code{t}.
}
\description{
Helper function necessary for calculating the log-likelihood.
}
