% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{integral_intensity}
\alias{integral_intensity}
\title{Computes term two of the log likelihood}
\usage{
integral_intensity(
  events,
  int_kernel,
  parameters,
  mu_fn = mu_none,
  mu_diff_fn = mu_diff_none,
  mu_int_fn = mu_int_none,
  print_level = 1
)
}
\arguments{
\item{events}{Vector of event times.}

\item{int_kernel}{Integral of kernel function for Hawkes Process.}

\item{parameters}{Parameters of the Hawkes kernel.}

\item{mu_fn}{Function that returns exogenous part of Hawkes Process.}

\item{mu_diff_fn}{Function that returns differential of exogenous part.}

\item{mu_int_fn}{Function that returns integral of exogenous part.}

\item{print_level}{Level at which logger will print}
}
\value{
Returns integral of the Hawkes intensity.
}
\description{
Term two of the log likelhood function is:
\deqn{\int_{0}^{T} \mu(\tau) d\tau - \sum_{i=1}^{n} \int_{t_{i}}^{T} g(\tau - t_{i} d\tau)}
}
