% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_intensities.R
\name{max_lambda}
\alias{max_lambda}
\title{Compute lambda_max - maximum intensity at a given time}
\usage{
max_lambda(
  time,
  events,
  previous_event_time,
  T_max,
  kernel,
  parameters,
  mu_fn = mu_none,
  mu_fn_diff = mu_diff_none,
  print_level = 1
)
}
\arguments{
\item{time}{Current time.}

\item{events}{Vector of event times.}

\item{previous_event_time}{Time of previous known event.}

\item{T_max}{Maximum time of simulation.}

\item{kernel}{Function describing the kernel.}

\item{parameters}{Parameters of the Hawkes kernel.}

\item{mu_fn}{Function describing contribution to intensity from exogenous terms.}

\item{mu_fn_diff}{Differential of the mu function.}

\item{print_level}{Level of printing to display.}
}
\value{
Maximum value of lambda after an event assuming no more events.
}
\description{
Compute lambda_max - maximum intensity at a given time
}
\examples{
max_lambda(time = 2, events = c(0.5, 1, 1.3), previous_event_time = 1.3,
           T_max = 10, kernel = ray_kernel,
           parameters = list("alpha" = 1.0, "delta" = 1.0, "delay" = 0))
max_lambda(time = 2, events = c(0.5, 1, 1.3), previous_event_time = 1.3,
           T_max = 10, kernel = ray_kernel,
           parameters = list("alpha" = 1.0, "delta" = 1.0, "delay" = 0, "A" = 2),
           mu_fn = mu_constant, mu_fn_diff = mu_diff_constant)
}
