% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_intensities.R
\name{max_mu}
\alias{max_mu}
\title{Compute time of maximum intensity of mu}
\usage{
max_mu(time, T_max, parameters, mu_fn = mu_none, mu_fn_diff = mu_diff_none)
}
\arguments{
\item{time}{Current time.}

\item{T_max}{Maximum time of simulation.}

\item{parameters}{Parameters of the Hawkes kernel.}

\item{mu_fn}{Function describing contribution to intensity from exogenous terms.}

\item{mu_fn_diff}{Differential of the mu function.}
}
\value{
Times of maximum value of mu.
}
\description{
Compute time of maximum intensity of mu
}
\examples{
max_mu(time = 5, T_max= 10, parameters = list("A" = 1),
       mu_fn = mu_constant, mu_fn_diff = mu_diff_constant)
max_mu(time = 5, T_max= 100, parameters = list("A" = 1, "B" = -1),
       mu_fn = mu_linear, mu_fn_diff = mu_diff_linear)
max_mu(time = 5, T_max= 700, parameters = list("M" = 1, "N" = 1),
      mu_fn = mu_sinusoidal, mu_fn_diff = mu_diff_sinusoidal)
}
