% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_constant}
\alias{mu_constant}
\title{Constant mu function}
\usage{
mu_constant(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A.}
}
\value{
Value of constant mu function.
}
\description{
Constant mu function
}
\examples{
mu_constant(time = 5, parameters = list("A" = 10))
mu_constant(time = 78, parameters = list("A" = 5))
}
