% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_diff_constant}
\alias{mu_diff_constant}
\title{Differential of constant mu function wrt time}
\usage{
mu_diff_constant(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A.}
}
\value{
Value of differential of constant mu function wrt time.
}
\description{
Differential of constant mu function wrt time
}
\examples{
mu_diff_constant(time = 5, parameters = list("A" = 10))
}
