% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_diff_linear}
\alias{mu_diff_linear}
\title{Differential of linear mu function wrt t}
\usage{
mu_diff_linear(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A and B.}
}
\value{
Value of differential of constant mu function wrt t.
}
\description{
Differential of linear mu function wrt t
}
\examples{
mu_diff_linear(time = 5, parameters = list("A" = 10, "B" = 4))
}
