% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_diff_none}
\alias{mu_diff_none}
\title{Differntial of no mu function wrt t}
\usage{
mu_diff_none(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.}
}
\value{
Value of differential of no mu function.
}
\description{
Differntial of no mu function wrt t
}
\examples{
mu_diff_none(time = 5, parameters = list("A" = 10))
}
