% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_diff_sinusoidal_linear_period}
\alias{mu_diff_sinusoidal_linear_period}
\title{Differential of linear and sinusoidal mu function wrt t with varying P}
\usage{
mu_diff_sinusoidal_linear_period(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A, B, M, N and P.}
}
\value{
Value of differential of linear and sinusoidal mu function with varying P.
}
\description{
Differential of linear and sinusoidal mu function wrt t with varying P
}
\examples{
mu_diff_sinusoidal_linear_period(time = 5, parameters = list("A" = 10, "B" = 1, "M" = 4,
                                                              "N" = 1, "P" = 365))
mu_diff_sinusoidal_linear_period(time = 78, parameters = list("A" = 5, "B" = 1, "M" = 4,
                                                               "N" = 1, "P" = 27))
}
