% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_int_constant}
\alias{mu_int_constant}
\title{Integral of constant mu function}
\usage{
mu_int_constant(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A.}
}
\value{
Value of integral of constant mu function.
}
\description{
Integral of constant mu function
}
\examples{
mu_int_constant(time = 5, parameters = list("A" = 10))
mu_int_constant(time = 78, parameters = list("A" = 5))
}
