% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_int_none}
\alias{mu_int_none}
\title{Integral of no mu function}
\usage{
mu_int_none(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.}
}
\value{
Value of integral of no mu function.
}
\description{
Integral of no mu function
}
\examples{
mu_int_none(time = 5, parameters = list("A" = 10))
}
