% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_int_quadratic}
\alias{mu_int_quadratic}
\title{Integral of quadratic mu function wrt t}
\usage{
mu_int_quadratic(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A, B and C.}
}
\value{
Value of integral of quadratic mu function wrt t.
}
\description{
Integral of quadratic mu function wrt t
}
\examples{
mu_int_quadratic(time = 5, parameters = list("A" = 10, "B" = 4, "C" = 1))
mu_int_quadratic(time = 78, parameters = list("A" = 5, "B" = 4, "C" = 1))
}
