% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_int_sinusoidal}
\alias{mu_int_sinusoidal}
\title{Integral of sinusoidal mu function}
\usage{
mu_int_sinusoidal(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define M and N.}
}
\value{
Value of integral of sinusoidal mu function.
}
\description{
Integral of sinusoidal mu function
}
\examples{
mu_int_sinusoidal(time = 5, parameters = list("M" = 4, "N" = 1))
mu_int_sinusoidal(time = 78, parameters = list("M" = 4, "N" = 1))
}
