% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_int_sinusoidal_constant}
\alias{mu_int_sinusoidal_constant}
\title{Integral of constant and sinusoidal mu function}
\usage{
mu_int_sinusoidal_constant(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A, M and N.}
}
\value{
Value of integral of constant and sinusoidal mu function.
}
\description{
Integral of constant and sinusoidal mu function
}
\examples{
mu_int_sinusoidal_constant(time = 5, parameters = list("A" = 10, "M" = 4, "N" = 1))
mu_int_sinusoidal_constant(time = 78, parameters = list("A" = 5, "M" = 4, "N" = 1))
}
