% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_linear}
\alias{mu_linear}
\title{Linear mu function}
\usage{
mu_linear(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A and B.}
}
\value{
Value of constant mu function.
}
\description{
Linear mu function
}
\examples{
mu_linear(time = 5, parameters = list("A" = 10, "B" = 4))
mu_linear(time = 78, parameters = list("A" = 5, "B" = 4))
}
