% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_none}
\alias{mu_none}
\title{No mu function}
\usage{
mu_none(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.}
}
\value{
Value of no mu function.
}
\description{
No mu function
}
\examples{
mu_none(time = 5, parameters = list("A" = 10))
}
