% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_quadratic}
\alias{mu_quadratic}
\title{Quadratic mu function}
\usage{
mu_quadratic(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define A, B and C.}
}
\value{
Value of quadratic mu function.
}
\description{
Quadratic mu function
}
\examples{
mu_quadratic(time = 5, parameters = list("A" = 10, "B" = 4, "C" = 1))
mu_quadratic(time = 78, parameters = list("A" = 5, "B" = 4, "C" = 1))
}
