% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{mu_sinusoidal}
\alias{mu_sinusoidal}
\title{Sinusoidal mu function}
\usage{
mu_sinusoidal(time, parameters)
}
\arguments{
\item{time}{Current time}

\item{parameters}{List of Hawkes Process parameters.  Need to define M and N.}
}
\value{
Value of sinusoidal mu function.
}
\description{
Sinusoidal mu function
}
\examples{
mu_sinusoidal(time = 5, parameters = list("M" = 4, "N" = 1))
mu_sinusoidal(time = 78, parameters = list("M" = 4, "N" = 1))
}
