% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{neg_log_likelihood}
\alias{neg_log_likelihood}
\title{Negative log likelihood}
\usage{
neg_log_likelihood(
  parameters,
  events,
  delay = 0,
  kernel,
  mu_fn = mu_none,
  mu_diff_fn = mu_diff_none,
  mu_int_fn = mu_int_none,
  print_level = 0
)
}
\arguments{
\item{parameters}{Parameters of the Hawkes kernel.}

\item{events}{Vector of event times.}

\item{delay}{Fixed delay}

\item{kernel}{Kernel function for Hawkes Process.}

\item{mu_fn}{Function that returns exogenous part of Hawkes Process.}

\item{mu_diff_fn}{Function that returns differential of exogenous part.}

\item{mu_int_fn}{Function that returns integral of exogenous part.}

\item{print_level}{Level at which logger will print}
}
\value{
Returns negative log-likelihood for optimising parameters.
}
\description{
The log likelihood is written as:
\deqn{log L = \sum_{i = 1}^{n} log (\mu(t_{i}) + \sum_{t_{i} - t_{j}} g(t_{i} - t_{j})) -
    \int_{0}^{T} \mu(t_{i}) d\tau - \sum_{i=1}^{n} \int_{t_{i}}^{T} g(\tau - t_{i}) d\tau}
     Therefore, the negative log likelihood is - log L.
}
\examples{
neg_log_likelihood(parameters = list("alpha" = 1, "delta" = 1, "delay" = 0),
    events = c(0, 1.5, 6, 9), kernel = exp_kernel)
neg_log_likelihood(parameters = list("alpha" = 1, "delta" = 1, "A" = 3, "delay" = 0),
    events = c(0, 2, 5, 9), kernel = ray_kernel,
    mu_fn = mu_constant, mu_int_fn = mu_int_constant)
}
