% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_decay_kernel}
\alias{plot_decay_kernel}
\title{Plot the decay kernel}
\usage{
plot_decay_kernel(kernel, parameters, T_max = 100)
}
\arguments{
\item{kernel}{Kernel.}

\item{parameters}{Parameters of the Hawkes kernel.}

\item{T_max}{Maximum length of kernel to consider}
}
\value{
Plot of decay kernel up to T_max.
}
\description{
Plot the decay kernel
}
\examples{
plot_decay_kernel(exp_kernel, parameters = list("alpha" = 1, "delta" = 1, "delay" = 0), T_max = 10)
plot_decay_kernel(ray_kernel, parameters = list("alpha" = 1, "delta" = 1, "delay" = 0), T_max = 10)
}
