% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_events}
\alias{plot_events}
\title{Plot events as counting process}
\usage{
plot_events(events, T_max, parent_indexes = NULL)
}
\arguments{
\item{events}{List of event times}

\item{T_max}{Maximum time to plot}

\item{parent_indexes}{Index of parent if known}
}
\value{
Plot of events as counting process
}
\description{
Plot events as counting process
}
\examples{
plot_events(events = c(0, 3, 6, 8.5), T_max = 10)
}
