% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_intensity_imported}
\alias{plot_intensity_imported}
\title{Plot force of infection for imported events}
\usage{
plot_intensity_imported(
  events,
  imported_events,
  T_max,
  kernel,
  parameters,
  mu_fn = mu_none,
  parent_indexes = NULL
)
}
\arguments{
\item{events}{List of event times}

\item{imported_events}{List of imported event times}

\item{T_max}{Maximum time to plot}

\item{kernel}{Kernel.}

\item{parameters}{Parameters of the Hawkes kernel}

\item{mu_fn}{Function for exogenous intensity (some examples are given in R/mu.R)}

\item{parent_indexes}{Index of parent if known}
}
\value{
Plot of force of infection.
}
\description{
Plot force of infection for imported events
}
\examples{
plot_intensity_imported(events = c(0, 2, 5, 6, 7), imported_events = c(2, 6), T_max = 10,
    kernel = exp_kernel, parameters = list("alpha" = 1, "delta" = 1, "delay" = 0))
plot_intensity_imported(events = c(0, 2, 5, 6, 7), imported_events = c(2, 6), T_max = 10,
    kernel = ray_kernel, parameters = list("alpha" = 1, "delta" = 1, "delay" = 0))
plot_intensity_imported(events = c(0, 2, 5, 6, 7), imported_events = c(2, 6), T_max = 10,
    kernel = ray_kernel, parameters = list("alpha" = 1, "delta" = 1, "delay" = 5))
}
