% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{print_message}
\alias{print_message}
\title{Logger function}
\usage{
print_message(msg, log_level, print_level)
}
\arguments{
\item{msg}{Message to be printed.}

\item{log_level}{Level of the logger at which this message should be printed.}

\item{print_level}{The level the code is running at}
}
\value{
Message printed to console.
}
\description{
Prints messages to screen at different levels of logging
}
