% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray_hessian.R
\name{ray_hessian}
\alias{ray_hessian}
\title{Computes hessian for Rayleigh kernel}
\usage{
ray_hessian(
  parameters,
  delay = 0,
  events,
  kernel,
  mu_fn = mu_none,
  mu_diff_fn = mu_diff_none,
  mu_int_fn = mu_int_none,
  print_level = 0
)
}
\arguments{
\item{parameters}{Parameters of the Hawkes kernel.}

\item{delay}{Fixed delay for kernel}

\item{events}{Vector of event times.}

\item{kernel}{Kernel function for Hawkes Process.}

\item{mu_fn}{Function that returns exogenous part of Hawkes Process.}

\item{mu_diff_fn}{Function that returns differential of exogenous part.}

\item{mu_int_fn}{Function that returns  integral of exogenous part.}

\item{print_level}{Level at which logger will print}
}
\value{
Returns directional derivatives for Rayleigh kernel.
}
\description{
Computes hessian for Rayleigh kernel
}
\examples{
ray_hessian(parameters = list("alpha" = 1, "delta" = 1), delay = 1,
    events = c(0, 4, 6, 9), kernel = exp_kernel, mu = mu_sinusoidal_linear,
    mu_diff_fn = mu_diff_sinusoidal_linear,
    mu_int_fn = mu_int_sinusoidal_linear)
}
