% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_next_event}
\alias{simulate_next_event}
\title{Simulate next event}
\usage{
simulate_next_event(
  time,
  events,
  T_max,
  num_children = 1,
  kernel,
  parameters,
  mu_fn = mu_none,
  mu_fn_diff = mu_diff_none,
  mu_t_max = NULL,
  imported = F,
  print_level = 1
)
}
\arguments{
\item{time}{Current time.}

\item{events}{List of previous events in simulation.}

\item{T_max}{Maximum time to simulate to.}

\item{num_children}{Number of children to simulate from each parent event.}

\item{kernel}{Kernel function}

\item{parameters}{Parameters for the Hawkes Kernel.}

\item{mu_fn}{Exogenous contribution to intensity.}

\item{mu_fn_diff}{Differential of exogenous contribution to intensity.}

\item{mu_t_max}{Maximum value of intensity contribution from exogenous terms.}

\item{imported}{Flag if simulating from just exogenous cases or all}

\item{print_level}{The level the code is running at}
}
\value{
Next event in Hawkes Process or a time after T_max if reached end of
    simulation period.
}
\description{
Uses Poisson thinning to select next event in Hawkes Process
}
\examples{
simulate_next_event(time = 0, events = c(0), T_max = 5,
    parameters = list("alpha" = 1, "delta" = 1, "delay" = 0),
    kernel = ray_kernel)
simulate_next_event(time = 5, events = c(0, 1.5, 6), T_max = 10,
    parameters = list("alpha" = 1, "delta" = 1, "delay" = 0),
    kernel = ray_kernel)
}
